Annotate Hotspots.

Annotate MAF-like data frome with a hot_spot column indicating recurrent mutations.

annotate_hotspots(
  mutation_maf,
  recurrence_min = 5,
  analysis_base = c("FL--DLBCL", "BL--DLBCL"),
  p_thresh = 0.05
)

Arguments

mutation_maf

A data frame in MAF format.

recurrence_min

minimum number of recurrences for mutation to be included, default is 5.

analysis_base

Base name for hot spot output directory.

p_thresh

P value threshold, default is 0.05.

Value

The same data frame with one additional column "hot_spot".

Details

This function takes an already loaded MAF data frame with the mutation_maf parameter. The user can then control the minimum number of recurrences for mutations to be included with recurrance_min, The default is 5. analysis_base controls the base name go hotspot output directory. Lastly, p_thresh sets the p value threshold, default is 0.05.

Examples

my_metadata = get_gambl_metadata()
all_coding_ssm = get_coding_ssm(these_samples_metadata = my_metadata,
                                projection = "grch37", 
                                seq_type = "genome")
#> reading from: /projects/nhl_meta_analysis_scratch/gambl/results_local/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/augmented_maf/all_slms-3--grch37.CDS.maf
#> mutations from 1652 samples
#> after linking with metadata, we have mutations from 1646 samples

hot_ssms = annotate_hotspots(all_coding_ssm)
#> Adding missing grouping variables: `SYMBOL`
#> Adding missing grouping variables: `SYMBOL`