Annotate IGH Breakpoints.

Add annotation to IGH breakpoints to infer mechanism based on location within IGH.

annotate_igh_breakpoints(annotated_df, genome_build = "grch37")

Arguments

annotated_df

Previously annotated data frame of SVs.

genome_build

Version of reference build to be used, only grch37 currently accepted.

Value

A slightly modified bedpe with added columns.

Details

Returns a modified bedpe with additional columns. The most useful columns that are added are mechanism (one of CSR, AID, VDJ) and label (NA if unmatched, otherwise one of Emu, Smu, one of the J segments or switch regions).

Examples

manta_sv = get_manta_sv(verbose = FALSE)
#> [1] "missing:  /Users/rmorin/gambl_results/gambl/svar_master-1.0/merges/gridss_manta.genome--grch37.bedpe"
#> Cannot find file locally. If working remotely, perhaps you forgot to load your config (see below) or sync your files?
#> Sys.setenv(R_CONFIG_ACTIVE = "remote")
#> Error: '/Users/rmorin/gambl_results/gambl/svar_master-1.0/merges/gridss_manta.genome--grch37.bedpe' does not exist.
all_annotated = annotate_sv(sv_data = manta_sv)
#> Error in dplyr::select(., "CHROM_A", "START_A", "END_A", "tumour_sample_id",     ends_with("SCORE"), "STRAND_A"): object 'manta_sv' not found
ig_annotated = annotate_igh_breakpoints(all_annotated)
#> Error in dplyr::filter(annotated_df, partner == "IGH"): object 'all_annotated' not found