Annotate SVs.

Annotate a data frame of SV breakpoints represented in an extended BEDPE format.

annotate_sv(
  sv_data,
  partner_bed,
  with_chr_prefix = FALSE,
  collapse_redundant = FALSE,
  return_as = "bedpe",
  blacklist = c(60565248, 30303126, 187728894, 101357565, 101359747, 161734970, 69400840,
    65217851, 187728889, 187728888, 187728892, 187728893, 188305164),
  genome_build = "grch37"
)

Arguments

sv_data

A data frame of SVs. This should be the output of get_manta_sv. If you aren't using the database backend you can supply your own data frame in the format show below. Most of this data is directly from the bedpe files that are obtained by converting the Manta outputs from VCF. Only the columns for both chromosomes, coordinates and strand plus SOMATIC_SCORE and tumour_sample_id are absolutely required. CHROM_A START_A END_A CHROM_B START_B END_B NAME SOMATIC_SCORE STRAND_A STRAND_B TYPE FILTER VAF_tumour VAF_normal DP_tumour DP_normal tumour_sample_id normal_sample_id pair_status 1 1556541 1556547 1 1556664 1556670 . 40 - - BND PASS 0.145 0 55 73 00-14595_tumorA 00-14595_normal matched

partner_bed

Optional bed-format data frame to use for annotating oncogene partners (e.g. enhancers). required columns are: chrom,start,end,gene

with_chr_prefix

Optionally request that chromosome names are returned with a chr prefix. Default is FALSE.

collapse_redundant

Remove reciprocal events and only return one per event. Default is FALSE.

return_as

Stated format for returned output, default is "bedpe". Other accepted output formats are "bed" and "bedpe_entrez" (to keep entrez_ids for compatibility with portal.R and cBioPortal).

blacklist

A vector of regions to be removed from annotations. Default coordinates are in respect to hg19.

genome_build

Reference genome build parameter, default is grch37.

Value

A data frame with annotated SVs (gene symbol and entrez ID).

Details

Specify a data frame with SVs (preferably the output from get_manta_sv) to the sv_df parameter and get back the same data frame with SV annotations.

Examples

sv_df = get_manta_sv(verbose = FALSE)
#> [1] "missing:  /Users/rmorin/gambl_results/gambl/svar_master-1.0/merges/gridss_manta.genome--grch37.bedpe"
#> Cannot find file locally. If working remotely, perhaps you forgot to load your config (see below) or sync your files?
#> Sys.setenv(R_CONFIG_ACTIVE = "remote")
#> Error: '/Users/rmorin/gambl_results/gambl/svar_master-1.0/merges/gridss_manta.genome--grch37.bedpe' does not exist.
annotated_entrez = annotate_sv(sv_data = sv_df,
                               with_chr_prefix = FALSE,
                               collapse_redundant = FALSE,
                               return_as = "bedpe_entrez", 
                               genome_build = "grch37")
#> Error in dplyr::select(., "CHROM_A", "START_A", "END_A", "tumour_sample_id",     ends_with("SCORE"), "STRAND_A"): object 'sv_df' not found