ASHM Multi-panel Rainbow Plot

Generates a colourful multi-panel overview of hypermutation in regions of interest across many samples.

ashm_multi_rainbow_plot(
  regions_bed,
  regions_to_display,
  exclude_classifications,
  metadata,
  seq_type,
  custom_colours,
  classification_column = "lymphgen",
  maf_data
)

Arguments

regions_bed

Bed file with chromosome coordinates, should contain columns chr, start, end, name (with these exact names).

regions_to_display

Optional vector of names from default regions_bed to use.

exclude_classifications

Optional argument for excluding specific classifications from a metadeta file.

metadata

A metadata file already subsetted and arranged on the order you want the samples vertically displayed.

seq_type

the seqtype you want results back for if maf_data is not provided.

custom_colours

Provide named vector (or named list of vectors) containing custom annotation colours if you do not want to use standartized pallette.

classification_column

Optional. Override default column for assigning the labels used for colouring in the figure.

maf_data

An already loaded maf, if no provided, this function will call get_ssm_by_region, using the regions supplied into regions_bed.

Value

Nothing

Details

The input for this function is a bed-file with the following columns; chr, start, end, name. Note that for this function to work, the column names must be exactly this. The user also needs to specify a vector of names (regions_to_display) to further control what regions are to be displayed on the returned plot. It is also possible to exclude specific classifications from the metadata file. This is achieved with exclude_classifications. In addition the user can also use the metadata parameter to use an already subset and arranged metadata table. This function will call get_ssm_by_region if maf_data is not called. For more info, refer to the parameter descriptions of this function.

Examples

#get lymphgen colours
lymphgen_colours = get_gambl_colours(classification = "lymphgen")

#build plot
ashm_multi_rainbow_plot(regions_to_display = c("BCL2-TSS",
                                               "MYC-TSS",
                                               "SGK1-TSS",
                                               "IGL"),
                        custom_colours = lymphgen_colours,
                        seq_type = "genome")
#> [1] "missing:  /Users/rmorin/git/gambl/data/metadata/gambl_all_outcomes.tsv"
#> Have you cloned the GAMBL repo and added the path to this directory under the local section of your config?
#> Error: '/Users/rmorin/git/gambl/data/metadata/gambl_all_outcomes.tsv' does not exist.