Make a rainbow plot of all mutations in a region, ordered and coloured by metadata

Make a rainbow plot of all mutations in a region, ordered and coloured by metadata

ashm_rainbow_plot(
  db = "gambl_test",
  table_name = "maf_slms3_hg19_icgc",
  mutations_maf,
  metadata,
  classification_column = "pathology",
  bed
)

Arguments

db
The GAMBL database name

table_name
The table we are querying

mutations_maf
A data frame containing mutations (MAF format) within a region of interest (i.e. use the get_ssm_by_region)

metadata
should be a data frame with sample_id as a column that should match Tumor_Sample_Barcode in the database

classification_column
The name of the metadata column to use for ordering and colouring samples

bed
Optional data frame specifying the regions to annotate (required columns: start, end, name)

Value

ggplot2 object

Examples

# basic usage
ashm_rainbow_plot(mutations_maf=my_mutations,metadata=my_metadata)

#> Error in select(., Tumor_Sample_Barcode, Start_Position): could not find function "select"

# advanced usages
mybed = data.frame(start=c(128806578,128805652,128748315), end=c(128806992,128809822,128748880), name=c("TSS","enhancer","MYC-e1"))
ashm_rainbow_plot(mutations_maf=my_mutations,metadata=my_metadata,bed=mybed)

#> Error in select(., Tumor_Sample_Barcode, Start_Position): could not find function "select"