Check GAMBLR Config.

Check that the GAMBLR config you have loaded will work in your setup.

check_gamblr_config(
  compare_timestamps = FALSE,
  ssh_session,
  archive_mode = FALSE,
  force_backup = FALSE
)

Arguments

compare_timestamps

Whether the function will compare timestamps on your local files to the remote copy. Only relevant if you are working remotely.

ssh_session

The ssh_session object see get_ssh_session() for more information. Only relevant if you are working remotely.

archive_mode

This is not currently working but the idea here is to keep a GSC archive of GAMBL in sync with the actively updated outputs

force_backup

Boolean parameter set to FALSE per default.

Details

This function is mostly for remote GAMBLRs to ensure they keep their local mirror of the GAMBL data up-to-date.

Examples

check_gamblr_config()
#> # A tibble: 4 × 2
#>   seq_type projection
#>   <chr>    <chr>     
#> 1 capture  grch37    
#> 2 capture  hg38      
#> 3 genome   grch37    
#> 4 genome   hg38      
#> Warning: argument is not an atomic vector; coercing
#> [1] "Multiple wildcards!"
#> [1] "seq_type"   "projection"
#> [1] "all_the_things/slms_3-1.0_vcf2maf-1.3/{seq_type}--projection/deblacklisted/maf/all_slms-3--{projection}.CDS.maf"
#> [1] "Multiple wildcards!"
#> [1] "seq_type"   "projection"
#> [1] "all_the_things/slms_3-1.0_vcf2maf-1.3/{seq_type}--projection/deblacklisted/augmented_maf/all_slms-3--{projection}.CDS.maf"
#> [1] "Multiple wildcards!"
#> [1] "seq_type"   "projection"
#> [1] "all_the_things/cnv_master-1.0/merges/{seq_type}--{projection}.seg"
#> Warning: argument is not an atomic vector; coercing
#> [1] "all_the_things/svar_master-1.0/merges/gridss_manta.genome--{projection}.bedpe"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/level_3/capture--grch37_clean_blacklist.txt
#> [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/level_3/capture--grch37_clean_blacklist.txt"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/level_3/capture--hg38_clean_blacklist.txt
#> [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/level_3/capture--hg38_clean_blacklist.txt"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/level_3/genome--grch37_clean_blacklist.txt
#> [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/level_3/genome--grch37_clean_blacklist.txt"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/level_3/genome--hg38_clean_blacklist.txt
#> [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/level_3/genome--hg38_clean_blacklist.txt"
#> Looking for: /Users/rmorin/gambl_results/shared/gambl_capture_results.tsv
#> [1] "/Users/rmorin/gambl_results/shared/gambl_capture_results.tsv"
#> Looking for: /Users/rmorin/gambl_results/shared/gambl_genome_results.tsv
#> [1] "/Users/rmorin/gambl_results/shared/gambl_genome_results.tsv"
#> Looking for: /Users/rmorin/gambl_results/icgc_dart/DESeq2-0.0_salmon-1.0/mrna--gambl-icgc-all/vst-matrix-Hugo_Symbol_tidy.tsv.gz
#> [1] "/Users/rmorin/gambl_results/icgc_dart/DESeq2-0.0_salmon-1.0/mrna--gambl-icgc-all/vst-matrix-Hugo_Symbol_tidy.tsv.gz"
#> Looking for: /Users/rmorin/gambl_results/gambl/DESeq2-0.0_salmon-1.0/mrna/vst-matrix-Hugo_Symbol_tidy.tsv
#> [1] "/Users/rmorin/gambl_results/gambl/DESeq2-0.0_salmon-1.0/mrna/vst-matrix-Hugo_Symbol_tidy.tsv"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/capture--projection/deblacklisted/maf/all_slms-3--grch37.CDS.maf
#> [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/capture--projection/deblacklisted/maf/all_slms-3--grch37.CDS.maf"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/capture--projection/deblacklisted/maf/all_slms-3--hg38.CDS.maf
#> [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/capture--projection/deblacklisted/maf/all_slms-3--hg38.CDS.maf"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/maf/all_slms-3--grch37.CDS.maf
#> [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/maf/all_slms-3--grch37.CDS.maf"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/maf/all_slms-3--hg38.CDS.maf
#> [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/maf/all_slms-3--hg38.CDS.maf"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/capture--projection/deblacklisted/augmented_maf/all_slms-3--grch37.CDS.maf
#> [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/capture--projection/deblacklisted/augmented_maf/all_slms-3--grch37.CDS.maf"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/capture--projection/deblacklisted/augmented_maf/all_slms-3--hg38.CDS.maf
#> [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/capture--projection/deblacklisted/augmented_maf/all_slms-3--hg38.CDS.maf"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/augmented_maf/all_slms-3--grch37.CDS.maf
#> [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/augmented_maf/all_slms-3--grch37.CDS.maf"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/augmented_maf/all_slms-3--hg38.CDS.maf
#> [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/augmented_maf/all_slms-3--hg38.CDS.maf"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/cnv_master-1.0/merges/capture--grch37.seg
#> [1] "/Users/rmorin/gambl_results/all_the_things/cnv_master-1.0/merges/capture--grch37.seg"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/cnv_master-1.0/merges/capture--hg38.seg
#> [1] "/Users/rmorin/gambl_results/all_the_things/cnv_master-1.0/merges/capture--hg38.seg"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/cnv_master-1.0/merges/genome--grch37.seg
#> [1] "/Users/rmorin/gambl_results/all_the_things/cnv_master-1.0/merges/genome--grch37.seg"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/cnv_master-1.0/merges/genome--hg38.seg
#> [1] "/Users/rmorin/gambl_results/all_the_things/cnv_master-1.0/merges/genome--hg38.seg"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/svar_master-1.0/merges/gridss_manta.genome--grch37.bedpe
#> [1] "/Users/rmorin/gambl_results/all_the_things/svar_master-1.0/merges/gridss_manta.genome--grch37.bedpe"
#> Looking for: /Users/rmorin/gambl_results/all_the_things/svar_master-1.0/merges/gridss_manta.genome--hg38.bedpe
#> [1] "/Users/rmorin/gambl_results/all_the_things/svar_master-1.0/merges/gridss_manta.genome--hg38.bedpe"
#> Warning: There were 22 files that cannot be found (see above). If this is unexpected, try to obtain them.
#> [1] "MISSING FILES:"
#>  [1] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/level_3/capture--grch37_clean_blacklist.txt"                               
#>  [2] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/level_3/capture--hg38_clean_blacklist.txt"                                 
#>  [3] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/level_3/genome--grch37_clean_blacklist.txt"                                
#>  [4] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/level_3/genome--hg38_clean_blacklist.txt"                                  
#>  [5] "/Users/rmorin/gambl_results/shared/gambl_capture_results.tsv"                                                                                
#>  [6] "/Users/rmorin/gambl_results/shared/gambl_genome_results.tsv"                                                                                 
#>  [7] "/Users/rmorin/gambl_results/icgc_dart/DESeq2-0.0_salmon-1.0/mrna--gambl-icgc-all/vst-matrix-Hugo_Symbol_tidy.tsv.gz"                         
#>  [8] "/Users/rmorin/gambl_results/gambl/DESeq2-0.0_salmon-1.0/mrna/vst-matrix-Hugo_Symbol_tidy.tsv"                                                
#>  [9] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/capture--projection/deblacklisted/maf/all_slms-3--grch37.CDS.maf"          
#> [10] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/capture--projection/deblacklisted/maf/all_slms-3--hg38.CDS.maf"            
#> [11] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/maf/all_slms-3--grch37.CDS.maf"           
#> [12] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/maf/all_slms-3--hg38.CDS.maf"             
#> [13] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/capture--projection/deblacklisted/augmented_maf/all_slms-3--grch37.CDS.maf"
#> [14] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/capture--projection/deblacklisted/augmented_maf/all_slms-3--hg38.CDS.maf"  
#> [15] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/augmented_maf/all_slms-3--grch37.CDS.maf" 
#> [16] "/Users/rmorin/gambl_results/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/augmented_maf/all_slms-3--hg38.CDS.maf"   
#> [17] "/Users/rmorin/gambl_results/all_the_things/cnv_master-1.0/merges/capture--grch37.seg"                                                        
#> [18] "/Users/rmorin/gambl_results/all_the_things/cnv_master-1.0/merges/capture--hg38.seg"                                                          
#> [19] "/Users/rmorin/gambl_results/all_the_things/cnv_master-1.0/merges/genome--grch37.seg"                                                         
#> [20] "/Users/rmorin/gambl_results/all_the_things/cnv_master-1.0/merges/genome--hg38.seg"                                                           
#> [21] "/Users/rmorin/gambl_results/all_the_things/svar_master-1.0/merges/gridss_manta.genome--grch37.bedpe"                                         
#> [22] "/Users/rmorin/gambl_results/all_the_things/svar_master-1.0/merges/gridss_manta.genome--hg38.bedpe"                                           
#> [1] "DONE!"