Compare segmented data for multiple samples.

cnvKompare returns a list in variable data formats allowing to evaluate concordance of CNV data between multiple samples.

cnvKompare(
  patient_id,
  these_sample_ids,
  this_seg,
  seg_path,
  genes_of_interest,
  projection = "grch37",
  this_seq_type = "genome",
  ignore_cytoband_labels = c("acen", "gvar", "stalk"),
  max_overlap = 20,
  min_concordance = 90,
  exclude_sex = FALSE,
  return_heatmap = TRUE,
  compare_pairwise = TRUE
)

Arguments

patient_id

Specify patient_id to retrieve sample ids from GAMBL metadata.

these_sample_ids

Optionally, specify sample ids for comparison.

this_seg

Optional input data frame of seg file. Must adhere to seg format.

seg_path

Optionally, specify the path to a local seg file. Must adhere to seg format.

genes_of_interest

Provide specific genes to be displayed on the time-series plot.

projection

Argument specifying the projection of seg file, which will determine coordinates of the cytobands. Default is grch37, but hg38 is also accepted.

this_seq_type

Seq type for returned CN segments. One of "genome" (default) or "capture".

ignore_cytoband_labels

Cytobands to be ignored. By default, "acen", "gvar", "stalk" are excluded.

max_overlap

For a time-series plot, how many maximum overlapping points are allowed?

min_concordance

Integer value from 0 to 100 to indicate the minimum required similarity between cytobands to be considered concordant. The default is 90 (90%).

exclude_sex

Boolean argument specifying whether to exclude sex chromosomes from calculation. Default is FALSE.

return_heatmap

Boolean argument specifying whether to return a heatmap of cnvKompare scores. Default is TRUE.

compare_pairwise

Boolean argument specifying whether to perform pairwise comparisons if there are more than 2 time points in the group. Default is TRUE.

Value

A list of overall and pairwise percent concordance, concordant and discordant cytobands, comparison heatmap of cnvKompare scores, and time series ggplot object.

Details

This function will compare CNV data between samples with multiple time points. It can also handle same-sample comparison between different CNV callers if the sample ID is specified in unique fashion. For groups with more than 2 samples, optionally pairwise comparisons can be performed. The comparison is made based on the internally calculated score, which reflects the percentage of each cytoband covered by CNV (rounded to the nearest 5%) and its absolute CN. Optionally, the heatmap of cnvKompare scores can be returned. In addition, the function will return all concordant and discordant cytobands. Finally, the time series plot of CNV log ratios will be returned for all lymphoma genes, with further functionality to subset it to a panel of genes of interest.

Examples

cnvKompare(patient_id = "13-26835",
           genes_of_interest = c("EZH2",
                                 "TP53",
                                 "MYC",
                                 "CREBBP",
                                 "GNA13"),
           projection = "hg38")
#> [1] "missing:  /Users/rmorin/git/gambl/data/metadata/gambl_all_outcomes.tsv"
#> Have you cloned the GAMBL repo and added the path to this directory under the local section of your config?
#> Error: '/Users/rmorin/git/gambl/data/metadata/gambl_all_outcomes.tsv' does not exist.