Collate Lymphgen.

Expand a sample_table (metadata) horizontally with different flavours of lymphgen data.

collate_lymphgen(
  these_samples_metadata,
  lymphgen_version = "default",
  tidy = FALSE
)

Arguments

these_samples_metadata

Optional parameter with metadata filtered for sample_ids of interest. If provided, this function will join lymphgen with this metadata, regardless of tidy TRUE/FALSE.

lymphgen_version

Version of selected lymphgen, default is "default".

tidy

Boolean parameter, set to TRUE for tidy format (i.e long format with no columns dropped). Default is FALSE, which returns the data in a wide format, keeping both the original Subtype. Prediction and tidied LymphGen values and puts the values from each "flavour" in its own column.

Value

A df with lymphgen information.

Details

This function takes a sample table (metadata) and adds different flavours of lymphgen data. It is possible to call this function with an already subset metadata table (with sample IDs of interest) with these_samples_metadata. If this is done, the function will join the lymphgen data with this table. Currently, the only supported lymphgen_version is "default". For more information refer to the function examples.

Examples

this_meta = get_gambl_metadata()
#> [1] "missing:  /Users/rmorin/git/gambl/data/metadata/gambl_all_outcomes.tsv"
#> Have you cloned the GAMBL repo and added the path to this directory under the local section of your config?
#> Error: '/Users/rmorin/git/gambl/data/metadata/gambl_all_outcomes.tsv' does not exist.
dlbcl_meta = dplyr::filter(this_meta, pathology == "DLBCL")
#> Error in dplyr::filter(this_meta, pathology == "DLBCL"): object 'this_meta' not found

wide_lymphgen = collate_lymphgen(these_samples_metadata = dlbcl_meta,
                                 lymphgen_version = "default",
                                 tidy = FALSE)
#> Error in mutate(df, `:=`({    {        lymphgen_column_out    }}, case_when(!str_detect(.data[[lymphgen_column_in]], "/") ~     .data[[lymphgen_column_in]], str_detect(.data[[lymphgen_column_in]],     "EZB") ~ "EZB-COMP", str_detect(.data[[lymphgen_column_in]],     "MCD") ~ "MCD-COMP", str_detect(.data[[lymphgen_column_in]],     "N1") ~ "N1-COMP", str_detect(.data[[lymphgen_column_in]],     "BN2") ~ "BN2-COMP", str_detect(.data[[lymphgen_column_in]],     "ST2") ~ "ST2-COMP"))): Problem while computing `LymphGen = case_when(...)`.
#> Caused by error in `.data[["Subtype.Prediction"]]`:
#> ! Column `Subtype.Prediction` not found in `.data`.