sample-level SV/SSM/CN reports in PDF

Construct pdf with sample-level plots, using minimum of arguments

comp_report(
  this_sample_id,
  export_individual_plots = FALSE,
  out,
  seq_data,
  seq_path = NULL,
  maf_data,
  maf_path = NULL,
  this_seq_type = "genome"
)

Arguments

this_sample_id

Sample ID to be plotted in report.

export_individual_plots

Boolean parameter, set to TRUE to export individual plots.

out

Path to output folder.

seq_data

Optional parameter with copy number df already loaded into R.

seq_path

Optional parameter with path to external cn file.

maf_data

Optional parameter with maf like df already loaded into R.

maf_path

Optional parameter with path to external maf like file.

this_seq_type

Seq type for returned CN segments. One of "genome" (default) or "capture".

Value

Nothing.

Details

This function runs the complete collection of fancy_x_plots for a specific sample ID (this_sample), with default parameters. The generated plots are put together into a two-page PDF. In addition, it is also possible to export all individual plots. This can be done by setting export_individual_plots to TRUE. It is also possible to use an already loaded seq file instead of using the this_sample_id parameter, this is done with the seq_data and maf_data parameters. Similarly, you can also point this function to a local file on disk with the seq_path and maf_path parameters.

Examples

if (FALSE) {
#create a PDF report for one sample, as well as exporting all individual plots.
comp_report(this_sample = "HTMCP-01-06-00422-01A-01D",
            out = "reports/",
            export_individual_plots = TRUE)
}