Custom cBioPortal case list.

Create and a custom case list for easy data subset in cBioPortal.

custom_caselist(
  these_sample_ids,
  caselist_name,
  caselist_description,
  return_missing_samples = FALSE,
  dir
)

Arguments

these_sample_ids

A vector of sample IDs to be subset into a case list. Required parameter.

caselist_name

Name of the generated case list (name does not include the file format, this will be added automagically). This parameter is required.

caselist_description

A verbose description of the created case list. Required.

return_missing_samples

Boolean parameter. Set to TRUE to return all sample IDs that are in the desired case list, but not represented in the study specific clinical file. Default is FALSE.

dir

The directory where all study specific files live.

Value

Nothing.

Details

Convenience function for specifying custom case lists that can be browsed on cBioPortal. This function takes a set of sample IDs these_sample_ids and intersect the IDs with what's available in the study-specific clinical file. This function also extracts the project name for the specified study, i.e the project name that is defined withing the folder specified under the dir parameter.

Examples

if (FALSE) {
#get some sample IDs
my_samples = get_gambl_metadata() %>%
 dplyr::filter(pathology == "FL") %>%
 dplyr::filter(cohort == "FL_GenomeCanada") %>%
 pull(sample_id)

#create case list with selected sample IDs
custom_caselist(these_sample_ids = my_samples,
                caselist_name = "FL_Canada",
                caselist_description = "Follicular Lymphoma from the Genome Canada Study",
                dir = "../path/to/study_directory/")
}