Estimate Purity.

Annotate a MAF with segmented absolute copy number data and added additional columns (VAF, Ploidy and Final_purity).

estimate_purity(
  in_maf,
  maf_df,
  in_seg,
  this_sample_id,
  this_seq_type = "genome",
  seg_file_source = "battenberg",
  show_plots = FALSE,
  assume_diploid = FALSE,
  coding_only = FALSE,
  genes
)

Arguments

in_maf

Path to a local maf file.

maf_df

Optional. Instead of using the path to a maf file, use a local dataframe as the maf file.

in_seg

Path to a local corresponding seg file for the same sample ID as the input maf.

this_sample_id

Specify the sample_id or any other string you want embedded in the file name.

this_seq_type

Seq type for returned CN segments. One of "genome" (default) or "capture".

seg_file_source

Specify what copy number calling program the input seg file is from, as it handles ichorCNA differently than WisecondorX, battenberg, etc.

show_plots

Optional. Show two faceted plots that display the VAF and purity distributions for each copy number state in the sample. Default is FALSE.

assume_diploid

Optional. If no local seg file is provided, instead of defaulting to a GAMBL sample, this parameter annotates every mutation as copy neutral. Default is FALSE.

coding_only

Optional. set to TRUE to restrict to only coding variants. Default is FALSE.

genes

Genes of interest.

Value

A list containing a data frame (MAF-like format) with the segmented absolute copy number data and three extra columns: VAF is the variant allele frequency calculated from the t_ref_count and t_alt_count Ploidy is the number of copies of an allele in the tumour cell Final_purity is the finalized purity estimation per mutation after considering different copy number states and LOH events.

Details

This function takes a sample ID with the this_sample_id parameter and calls assign_cn_to_ssm to get CN information. The user can also use an already loaded maf file with maf_df. In addition, a path to the maf/seq file of interest can also be passed to this function with in_maf and in_seg. To visualize VAF and purity distributions, set the show_plots to TRUE (default is FALSE). For more information on how to run this function with the parameters at hand, refer to the parameter descriptions and function examples.

Examples

#load a maf
this_maf = get_ssm_by_sample(this_sample_id = "HTMCP-01-06-00422-01A-01D",
                             this_seq_type = "genome")
#> [1] "missing:  /Users/rmorin/git/gambl/data/metadata/gambl_all_outcomes.tsv"
#> Have you cloned the GAMBL repo and added the path to this directory under the local section of your config?
#> Error: '/Users/rmorin/git/gambl/data/metadata/gambl_all_outcomes.tsv' does not exist.

#estimate purity based on an already loaded maf object
estimate_purity(maf_df = this_maf,
                show_plots = TRUE)
#> Error in dplyr::mutate(., Chromosome = gsub("chr", "", Chromosome)): object 'this_maf' not found

#estimate purity based sole on a smaple ID + added seg data.
estimate_purity(this_sample_id = "HTMCP-01-06-00422-01A-01D",
                show_plots = TRUE, coding_only = TRUE)
#> [1] "missing:  /Users/rmorin/git/gambl/data/metadata/gambl_all_outcomes.tsv"
#> Have you cloned the GAMBL repo and added the path to this directory under the local section of your config?
#> Error: '/Users/rmorin/git/gambl/data/metadata/gambl_all_outcomes.tsv' does not exist.