Genome-wide ideogram (CN segments) for multiple samples.

Generate ideograms for selected samples, visualizing copy number variation segments.

fancy_multisamp_ideogram(
  these_sample_ids,
  plot_title = "CN Segments Ideogram",
  plot_sub = "grch37",
  this_seq_type = "genome",
  chr_anno_dist = 3,
  chr_select = paste0("chr", c(1:22)),
  include_cn2 = FALSE,
  kompare = FALSE,
  concordance = TRUE,
  coding_only = FALSE,
  from_flatfile = TRUE,
  use_augmented_maf = TRUE
)

Arguments

these_sample_ids

Sample to be plotted (accepts 2, 3 or 4 samples).

plot_title

Main title of plot.

plot_sub

Subtitle of plot.

this_seq_type

Seq type for returned CN segments. One of "genome" (default) or "capture".

chr_anno_dist

Optional parameter to adjust chromosome annotations, default value is 3, increase to adjust annotations to left.

chr_select

Optional parameter to subset plot to specific chromosomes. Default value is chr1-22.

include_cn2

Set to TRUE for plotting CN states == 2.

kompare

Boolean statement, set to TRUE to call cnvKompare on the selected samples for plotting concordant (or discordant) cn segments across selected chromosomes.

concordance

Boolean parameter to be used when kompare = TRUE. Default is TRUE, to plot discordant segments, set the parameter to FALSE.

coding_only

Optional. Set to TRUE to restrict to plotting only coding mutations.

from_flatfile

If set to true the function will use flat files instead of the database.

use_augmented_maf

Boolean statement if to use augmented maf, default is FALSE.

Value

A plot as a ggplot object (grob).

Details

To create multi-sample ideograms, i.e showing CN segments across multiple samples, this function was created. This can be used to infer inheritance patterns, hotspots, etc. across multiple samples or the sample ID but for different timepoints. In addition, this plot can also allow to only plot concordant (or discordant) cn segments between two samples. i.e how two samples differ, or are alike. The function automatically detects the number of samples provided and sets the plotting parameters accordingly. The maximum number of samples this plot can deal with is 4 and sample IDs of interest are given to the these_sample_ids. In order to only plot segments that are concordant between the selected samples, set komapre to TRUE and concordance to TRUE. To instead plot discordant CN segments, set this parameter to FALSE.

Examples

#two samples ideogram
fancy_multisamp_ideogram(these_sample_ids = c("00-15201_tumorA",
                                              "00-15201_tumorB"),
                         include_cn2 = TRUE,
                         plot_title = "Multi-sample Ideograms Example",
                         plot_sub = "grch37",
                         chr_anno_dist = 2.5,
                         chr_select = paste0("chr", c(1:22)))
#> restricting to non-ICGC data
#> [1] "missing:  /Users/rmorin/gambl_results/gambl/cnv_master-1.0/99-outputs/genome--projection/all--grch37.seg"
#> Cannot find file locally. If working remotely, perhaps you forgot to load your config (see below) or sync your files?
#> Sys.setenv(R_CONFIG_ACTIVE = "remote")
#> Error: '/Users/rmorin/gambl_results/gambl/cnv_master-1.0/99-outputs/genome--projection/all--grch37.seg' does not exist.