Retrieve the nearest SV to a specified SV in a given patient

Retrieve the nearest SV to a specified SV in a given patient

fetch_next_sv(
  db = "gambl_test",
  table_name = "bedpe_manta_hg19",
  min_vaf = 0.1,
  min_score = 40,
  pass = TRUE,
  sv_data,
  with_chr_prefix = FALSE,
  in_from
)

Arguments

db
The GAMBL database name

table_name
The table we are querying

min_vaf
The minimum tumour VAF for a SV to be returned

min_score
The lowest Manta somatic score for a SV to be returned

sv_data
A single row of a data frame returned by a function such as get_manta_sv

with_chr_prefix
Prepend all chromosome names with chr (required by some downstream analyses)

pair_status
Use to restrict results (if desired) to matched or unmatched results (default is to return all)

Value

A data frame in a bedpe-like format with additional columns that allow filtering of high-confidence SVs