Finalize Study (cBioPortal).

Finish setting up a new cBioPortal instance or updating an existing portal data set.

finalize_study(
  seq_type_filter = "genome",
  short_name = "GAMBL",
  human_friendly_name = "GAMBL data",
  project_name = "gambl_genome",
  description = "GAMBL data from genome",
  cancer_type = "mixed",
  these_sample_ids,
  overwrite = TRUE,
  meta_columns,
  out_dir
)

Arguments

seq_type_filter

the seq type you are setting up a study for, default is "genome".

short_name

A concise name for your portal project.

human_friendly_name

A slightly more verbose name for your project.

project_name

Unique ID for your project.

description

A verbose description of your data set.

cancer_type

Cancer types included in study, default is "mixed".

these_sample_ids

A vector of all the sample_id that were included in any of the data files for cBioPortal (i.e the output from setup_study and setup_fusions).

overwrite

Flag to specify that files should be overwritten if they exist. Default is TRUE.

meta_columns

Optional parameter for specifying metadata fields in the study. If not provided, the function will resort to using default metadata columns.

out_dir

The full path to the base directory where the files are being created.

Value

Nothing.

Details

This function should be run as the last (or third step) in setting up a new cBioPortal instance. The functions that should be run prior to these functions are; setup_study and setup_fusions. finalize_study creates all the necessary tables and metadata files (case lists) that are required to import a new study into cBioPortal. Note, that all parameter arguments used in this function have to match the same parameter arguments for the previously run functions (setup_study and setup_fusions).

Examples

if (FALSE) {
finalize_study(these_sample_ids = c(ids, fusion_ids), out_dir = "GAMBLR/cBioPortal/instance01/")
}