Get Combined SV.

Retrieve Combined Manta and GRIDSS-derived SVs from a flatfile and filter.

get_combined_sv(
  min_vaf = 0,
  these_sample_ids,
  with_chr_prefix = FALSE,
  projection = "grch37",
  oncogenes
)

Arguments

min_vaf

The minimum tumour VAF for a SV to be returned. Recommended: 0. (default: 0)

these_sample_ids

A character vector of tumour sample IDs you wish to retrieve SVs for.

with_chr_prefix

Prepend all chromosome names with chr (required by some downstream analyses). Default is FALSE.

projection

The projection genome build. Default is "grch37".

oncogenes

A character vector of genes commonly involved in translocations. Possible values: CCND1, CIITA, SOCS1, BCL2, RFTN1, BCL6, MYC, PAX5.

Value

A data frame in a bedpe-like format with additional columns that allow filtering of high-confidence SVs.

Details

The bedpe files used as input to this function were pre-filtered for a minimum VAF of 0.05, and SVs affecting. common translocation regions (BCL2, BCL6, MYC, CCND1) were whitelisted (e.g. no VAF filter applied). Therefore if you wish to post-filter the SVs we recommend doing so carefully after loading this data frame. Further, the input bedpe file is annotated with oncogenes and superenhancers from naive and germinal centre B-cells. You can subset to events affecting certain loci using the "oncogenes" argument. Is this function not what you are looking for? Try one of the following, similar, functions; get_manta_sv, get_manta_sv_by_sample, get_manta_sv_by_samples

Examples

get_combined_sv(oncogenes = c("MYC", "BCL2", "BCL6"))
#> # A tibble: 17,573 × 20
#>    CHROM_A   START_A     END_A CHROM_B  START_B  END_B manta_name SCORE STRAND_A
#>    <chr>       <dbl>     <dbl> <chr>      <dbl>  <dbl> <chr>      <dbl> <chr>   
#>  1 14      106325575 106325579 8         1.29e8 1.29e8 MantaBND:…    NA +       
#>  2 14      106325996 106325998 8         1.29e8 1.29e8 MantaBND:…    NA -       
#>  3 14      106211834 106211836 8         1.29e8 1.29e8 MantaBND:…    NA -       
#>  4 14      106324104 106324108 8         1.29e8 1.29e8 MantaBND:…    NA +       
#>  5 14      106326646 106326649 8         1.29e8 1.29e8 MantaBND:…    NA -       
#>  6 14      106330376 106330685 8         1.29e8 1.29e8 MantaBND:…    NA +       
#>  7 14      106329932 106329935 8         1.29e8 1.29e8 MantaBND:…    NA +       
#>  8 14      106329938 106329940 8         1.29e8 1.29e8 MantaBND:…    NA -       
#>  9 18       60368230  60368530 3         1.83e8 1.83e8 MantaBND:…    NA +       
#> 10 22       23247400  23247402 8         1.29e8 1.29e8 MantaBND:…    NA -       
#> # ℹ 17,563 more rows
#> # ℹ 11 more variables: STRAND_B <chr>, tumour_sample_id <chr>,
#> #   normal_sample_id <chr>, VAF_tumour <dbl>, DP <dbl>, gridss_name <chr>,
#> #   ANNOTATION_A <chr>, DIST_TO_ANNOTATION_A <dbl>, ANNOTATION_B <chr>,
#> #   DIST_TO_ANNOTATION_B <dbl>, FILTER <chr>