Get GAMBL Colours.

Get GAMBL colour schemes for annotating figures.

get_gambl_colours(
  classification = "all",
  alpha = 1,
  as_list = FALSE,
  as_dataframe = FALSE,
  return_available = FALSE,
  verbose = FALSE
)

Arguments

classification

Optionally request only colours for pathology, lymphgen, mutation or copy_number.

alpha

Alpha of plotted colours.

as_list

Boolean parameter controlling the format of the return. Default is FALSE.

as_dataframe

Boolean parameter controlling the format of the return. Default is FALSE.

return_available

Set to TRUE for returning all available colours. Default is FALSE.

verbose

Default is FALSE

Value

A named vector of colour codes for lymphgen classes and pathology.

Details

This function was designed to retrieve specified GAMBL colour palettes. By default, this function returns all the colours currently available. The user can easily specify what classification to return colors for with the classification parameter. It is also possible to return any given colour in different formats. To do so, refer to the Boolean arguments; as_list and as_dataframe. For more information regarding the available colours, refer to the utilities vignette.

Examples

lymphgen_cols = get_gambl_colours("lymphgen")

if (FALSE) {
#be sure to install ggsci from https://github.com/morinlab/ggsci
#install_github("morinlab/ggsci")
}