Retrieve Manta SVs from the database and filter

Retrieve Manta SVs from the database and filter

get_manta_sv(
  db = "gambl_test",
  table_name = "bedpe_manta_hg19",
  min_vaf = 0.1,
  min_score = 40,
  pass = TRUE,
  pairing_status,
  sample_id,
  chromosome,
  qstart,
  qend,
  region,
  with_chr_prefix = FALSE
)

Arguments

db
The GAMBL database name

table_name
The table we are querying

min_vaf
The minimum tumour VAF for a SV to be returned

min_score
The lowest Manta somatic score for a SV to be returned

chromosome
The chromosome you are restricting to

qstart
Query start coordinate of the range you are restricting to

qend
Query end coordinate of the range you are restricting to

region
Region formatted like chrX:1234-5678 instead of specifying chromosome, start and end separately

with_chr_prefix
Prepend all chromosome names with chr (required by some downstream analyses)

pair_status
Use to restrict results (if desired) to matched or unmatched results (default is to return all)

Value

A data frame in a bedpe-like format with additional columns that allow filtering of high-confidence SVs

Examples

#lazily get every SV in the table with default quality filters
all_sv = get_manta_sv()

#> Error in tbl(con, table_name) %>% filter(VAF_tumour >= min_vaf & SOMATIC_SCORE >=     min_score): could not find function "%>%"

#get all SVs for a single sample
some_sv = get_manta_sv(sample_id="94-15772_tumorA")

#> Error in tbl(con, table_name) %>% filter(VAF_tumour >= min_vaf & SOMATIC_SCORE >=     min_score): could not find function "%>%"

#get the SVs in a region around MYC
myc_locus_sv = get_manta_sv(region="8:128723128-128774067")

#> Error in tbl(con, table_name) %>% filter((CHROM_A == chromosome & START_A >=     qstart & START_A <= qend) | (CHROM_B == chromosome & START_B >=     qstart & START_B <= qend)) %>% filter(VAF_tumour >= min_vaf &     SOMATIC_SCORE >= min_score): could not find function "%>%"