Get Manta SV By Samples.

Load the manta output for a set of samples.

get_manta_sv_by_samples(
  these_samples_metadata,
  min_vaf = 0.1,
  min_score = 40,
  pass = TRUE,
  projection = "grch37",
  verbose = TRUE
)

Arguments

these_samples_metadata

The only required parameter is a metadata table (data frame) that must contain a row for each sample you want the data from. The additional columns the data frame needs to contain, besides sample_id, are: unix_group, genome_build, seq_type, pairing_status.

min_vaf

The minimum tumour VAF for a SV to be returned. Default value is 0.1.

min_score

The lowest Manta somatic score for a SV to be returned. Default value is 40.

pass

If set to TRUE, only return SVs that are annotated with PASS in the FILTER column. Set to FALSE to keep all variants, regardless if they PASS the filters. Default is TRUE.

projection

The projection of returned calls. Default is grch37.

verbose

Set to FALSE to prevent the path of the requested bedpe file to be printed.

Value

A data frame containing the Manta outputs from all sample_id in these_samples_metadata in a bedpe-like format with additional columns extracted from the VCF column.

Details

This is a convenience wrapper function for get_manta_sv_by_sample (and called by get_manta_sv). Is this function not what you are looking for? Try one of the following, similar, functions; get_combined_sv, get_manta_sv, get_manta_sv_by_sample

Examples

all_sv = get_manta_sv()
#> Reading 01-16433_tumorC from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/01-16433_tumorC--01-16433_normal--matched.somaticSV.bedpe
#> Reading 012-02-1TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/012-02-1TD--012-0034-01ND--matched.somaticSV.bedpe
#> Reading 012-11-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/012-11-01TD--012-0034-01ND--matched.somaticSV.bedpe
#> Reading 012-13-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/012-13-01TD--012-0034-01ND--matched.somaticSV.bedpe
#> Reading 012-15-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/012-15-01TD--012-0034-01ND--matched.somaticSV.bedpe
#> Reading 012-16-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/012-16-01TD--012-0034-01ND--matched.somaticSV.bedpe
#> Reading 012-17-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/012-17-01TD--012-0034-01ND--matched.somaticSV.bedpe
#> Reading 012-18-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/012-18-01TD--012-0034-01ND--matched.somaticSV.bedpe
#> Reading 012-19-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/012-19-01TD--012-0034-01ND--matched.somaticSV.bedpe
#> Reading 019-0047-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/019-0047-01TD--019-0048-01ND--matched.somaticSV.bedpe
#> Reading 019-05-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/019-05-01TD--019-0048-01ND--matched.somaticSV.bedpe
#> Reading 019-12-02TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/019-12-02TD--019-0048-01ND--matched.somaticSV.bedpe
#> Reading 019-15-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/019-15-01TD--019-0048-01ND--matched.somaticSV.bedpe
#> Reading 019-18-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/019-18-01TD--019-0048-01ND--matched.somaticSV.bedpe
#> Reading 019-19-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/019-19-01TD--019-0048-01ND--matched.somaticSV.bedpe
#> Reading 02-11368_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/02-11368_tumorB--02-11368_normal--matched.somaticSV.bedpe
#> Reading 02-15745_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/02-15745_tumorA--02-15745_normal--matched.somaticSV.bedpe
#> Reading 02-15745_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/02-15745_tumorB--02-15745_normal--matched.somaticSV.bedpe
#> Reading 02-15745_tumorC from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/02-15745_tumorC--02-15745_normal--matched.somaticSV.bedpe
#> Reading 03-24364_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/03-24364_tumorA--03-24364_normal--matched.somaticSV.bedpe
#> Reading 03-28283_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/03-28283_tumorB--03-28283_normal--matched.somaticSV.bedpe
#> Reading 05-11115_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/05-11115_tumorA--05-11115_normal--matched.somaticSV.bedpe
#> Reading 05-11115_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/05-11115_tumorB--05-11115_normal--matched.somaticSV.bedpe
#> Reading 05-11115_tumorC from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/05-11115_tumorC--05-11115_normal--matched.somaticSV.bedpe
#> Reading 05-11115_tumorD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/05-11115_tumorD--05-11115_normal--matched.somaticSV.bedpe
#> Reading 05-11115_tumorE from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/05-11115_tumorE--05-11115_normal--matched.somaticSV.bedpe
#> Reading 05-11725_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/05-11725_tumorB--05-11725_normal--matched.somaticSV.bedpe
#> Reading 05-15635_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/05-15635_tumorA--05-15635_normal--matched.somaticSV.bedpe
#> Reading 05-15635_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/05-15635_tumorB--05-15635_normal--matched.somaticSV.bedpe
#> Reading 05-27675_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/05-27675_tumorA--05-27675_normal--matched.somaticSV.bedpe
#> Reading 05-28753_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/05-28753_tumorA--05-28753N--matched.somaticSV.bedpe
#> Reading 05-28753_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/05-28753_tumorB--05-28753N--matched.somaticSV.bedpe
#> Reading 05-28753_tumorC from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/05-28753_tumorC--05-28753N--matched.somaticSV.bedpe
#> Reading 06-28477_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/06-28477_tumorA--FL2002N--matched.somaticSV.bedpe
#> Reading 063-0127-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/063-0127-01TD--063-0126-02ND--matched.somaticSV.bedpe
#> Reading 063-02-1TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/063-02-1TD--063-0126-02ND--matched.somaticSV.bedpe
#> Reading 063-08-04BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/063-08-04BD--063-0126-02ND--matched.somaticSV.bedpe
#> Reading 063-09-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/063-09-01TD--063-0126-02ND--matched.somaticSV.bedpe
#> Reading 08-26563_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/08-26563_tumorA--08-26563_normal--matched.somaticSV.bedpe
#> Reading 09-15148_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/09-15148_tumorB--09-15148_normal--matched.somaticSV.bedpe
#> Reading 09-31601_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/09-31601_tumorB--09-31601_normal--matched.somaticSV.bedpe
#> Reading 10-13016_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/10-13016_tumorA--10-13016_normal--matched.somaticSV.bedpe
#> Reading 10-18191T from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/10-18191T--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 102-01-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/102-01-01TD--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 102-0202-1DVT from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/102-0202-1DVT--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 11-13204_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/11-13204_tumorB--11-13204_normal--matched.somaticSV.bedpe
#> Reading 12-11721_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/12-11721_tumorB--12-11721_normal--matched.somaticSV.bedpe
#> Reading 12-38161_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/12-38161_tumorA--12-38161_normal--matched.somaticSV.bedpe
#> Reading 13-27960_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/13-27960_tumorB--13-27960N--matched.somaticSV.bedpe
#> Reading 13-27975_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/13-27975_tumorA--13-27975_normal--matched.somaticSV.bedpe
#> Reading 13-27975_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/13-27975_tumorB--13-27975_normal--matched.somaticSV.bedpe
#> Reading 13F44-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/13F44-PL02--13089-DNA01--matched.somaticSV.bedpe
#> Reading 13I80-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/13I80-PL02--13575-DNA01--matched.somaticSV.bedpe
#> Reading 13N66-PL01 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/13N66-PL01--13G42-DNAGL01--matched.somaticSV.bedpe
#> Reading 14-11465_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/14-11465_tumorB--14-11465_normal--matched.somaticSV.bedpe
#> Reading 14-11777_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/14-11777_tumorB--14-11777N--matched.somaticSV.bedpe
#> Reading 14-26768_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/14-26768_tumorA--14-26768_normal--matched.somaticSV.bedpe
#> Reading 140127-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/140127-PL02--13430-GL-DNA--matched.somaticSV.bedpe
#> Reading 141498-PL01 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/141498-PL01--13C53-DNAGL02--matched.somaticSV.bedpe
#> Reading 141900-PL01 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/141900-PL01--141666-BC01--matched.somaticSV.bedpe
#> Reading 150327-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/150327-PL02--13F62-BC01--matched.somaticSV.bedpe
#> Reading 1523-01-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/1523-01-01TD--1523-04-01ND--matched.somaticSV.bedpe
#> Reading 1523-03-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/1523-03-01BD--1523-04-01ND--matched.somaticSV.bedpe
#> Reading 1563-01-03TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/1563-01-03TD--1563-01-01ND--matched.somaticSV.bedpe
#> Reading 1563-02-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/1563-02-01BD--1563-01-01ND--matched.somaticSV.bedpe
#> Reading 160745-PL04 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/160745-PL04--160942-BC01--matched.somaticSV.bedpe
#> Reading 161135-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/161135-PL02--160942-BC01--matched.somaticSV.bedpe
#> Reading 161396-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/161396-PL02--160942-BC01--matched.somaticSV.bedpe
#> Reading 161402-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/161402-PL02--140928-BC01--matched.somaticSV.bedpe
#> Reading 1669-04-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/1669-04-01BD--1669-02-01ND--matched.somaticSV.bedpe
#> Reading 170867-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/170867-PL02--170458-BC01--matched.somaticSV.bedpe
#> Reading 171116-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/171116-PL02--170484-CS01--matched.somaticSV.bedpe
#> WARNING! No SV calls found in flat-file for: 171116-PL02
#> Reading 171290-PL01 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/171290-PL01--170458-BC01--matched.somaticSV.bedpe
#> Reading 171414-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/171414-PL02--171241-BC01--matched.somaticSV.bedpe
#> Reading 171447-PL01 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/171447-PL01--190142-DNAGL02--matched.somaticSV.bedpe
#> WARNING! No SV calls found in flat-file for: 171447-PL01
#> Reading 171451-PL01 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/171451-PL01--171328-BC01--matched.somaticSV.bedpe
#> WARNING! No SV calls found in flat-file for: 171451-PL01
#> Reading 180204-PL01-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/180204-PL01-PL02--171364-BC01--matched.somaticSV.bedpe
#> Reading 180445-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/180445-PL02--171241-BC01--matched.somaticSV.bedpe
#> Reading 180840-PL01 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/180840-PL01--180840-CS01--matched.somaticSV.bedpe
#> Reading 181712-PL01 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/181712-PL01--181677-BC01--matched.somaticSV.bedpe
#> Reading 190017-PL01 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/190017-PL01--181510-CS01--matched.somaticSV.bedpe
#> Reading 190128-PL01-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/190128-PL01-PL02--161134-DNAGL02--matched.somaticSV.bedpe
#> Reading 190806-PL01-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/190806-PL01-PL02--141055-DNAGL02--matched.somaticSV.bedpe
#> Reading 191457-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/191457-PL02--190142-DNAGL02--matched.somaticSV.bedpe
#> Reading 191510-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/191510-PL02--190263-BC01--matched.somaticSV.bedpe
#> Reading 191621-PL01 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/191621-PL01--190906-CS01--matched.somaticSV.bedpe
#> Reading 200836-PL01 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/200836-PL01--201151-BC01--matched.somaticSV.bedpe
#> Reading 210541-PL02 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hg38/somaticSV/210541-PL02--211147-CS01--matched.somaticSV.bedpe
#> Reading 3034-03-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/3034-03-01TD--3034-01-01ND--matched.somaticSV.bedpe
#> Reading 3034-04-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/3034-04-01BD--3034-01-01ND--matched.somaticSV.bedpe
#> Reading 3299-01-02TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/3299-01-02TD--3299-02-03ND--matched.somaticSV.bedpe
#> Reading 3299-02-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/3299-02-01TD--3299-02-03ND--matched.somaticSV.bedpe
#> Reading 3299-03-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/3299-03-01TD--3299-02-03ND--matched.somaticSV.bedpe
#> Reading 3299-04-1TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/3299-04-1TD--3299-02-03ND--matched.somaticSV.bedpe
#> Reading 3299-05-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/3299-05-01TD--3299-02-03ND--matched.somaticSV.bedpe
#> Reading 3495-02-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/3495-02-01BD--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 3495-03-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/3495-03-01BD--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 365-11-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/365-11-01BD--365-01-01ND--matched.somaticSV.bedpe
#> Reading 365-16-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/365-16-01TD--365-01-01ND--matched.somaticSV.bedpe
#> Reading 365-1951-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/365-1951-01TD--365-01-01ND--matched.somaticSV.bedpe
#> Reading 4675-03-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/4675-03-01BD--4675-05-01ND--matched.somaticSV.bedpe
#> Reading 4675-04-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/4675-04-01BD--4675-05-01ND--matched.somaticSV.bedpe
#> Reading 4676-02-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/4676-02-01BD--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 4686-02-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/4686-02-01TD--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 4686-03-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/4686-03-01BD--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 4687-02-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/4687-02-01TD--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 4687-03-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/4687-03-01BD--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 4690-02-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/4690-02-01TD--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 4690-03-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/4690-03-01BD--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 816-01-1TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/816-01-1TD--816-05-02ND--matched.somaticSV.bedpe
#> Reading 816-06-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/816-06-01TD--816-05-02ND--matched.somaticSV.bedpe
#> Reading 816-08-01TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/816-08-01TD--816-05-02ND--matched.somaticSV.bedpe
#> Reading 835-01-02BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/835-01-02BD--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 835-02-03TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/835-02-03TD--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 839-01-1TD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/839-01-1TD--839-03-1ND--matched.somaticSV.bedpe
#> Reading 839-04-01BD from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/839-04-01BD--839-03-1ND--matched.somaticSV.bedpe
#> Reading 97-28459_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/97-28459_tumorB--FL3006N--matched.somaticSV.bedpe
#> Reading 98-28290T from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/98-28290T--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 98-30186_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/98-30186_tumorA--98-30186_normal--matched.somaticSV.bedpe
#> Reading 99-25549T from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--hs37d5/somaticSV/99-25549T--14-11247Normal--unmatched.somaticSV.bedpe
#> Reading 99-27783_tumorA from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/99-27783_tumorA--99-27783_normal--matched.somaticSV.bedpe
#> Reading 99-27783_tumorB from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/99-27783_tumorB--99-27783_normal--matched.somaticSV.bedpe
#> Reading KM-H2 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/KM-H2--14-11247N--unmatched.somaticSV.bedpe
#> Reading L-428 from: /projects/nhl_meta_analysis_scratch/gambl/results_local/gambl/manta_current/99-outputs/bedpe/genome--grch37/somaticSV/L-428--14-11247N--unmatched.somaticSV.bedpe
metadata = get_gambl_metadata()
missing_samples = dplyr::anti_join(metadata, all_sv, by = c("sample_id" = "tumour_sample_id"))

missing_from_merge = get_manta_sv_by_samples(these_samples_metadata = missing_samples, verbose = FALSE)
#> WARNING! No SV calls found in flat-file for: 171116-PL02
#> WARNING! No SV calls found in flat-file for: 171447-PL01
#> WARNING! No SV calls found in flat-file for: 171451-PL01