GetSample CN Segments.

Get all segments for a single (or multiple) sample_id(s).

get_sample_cn_segments(
  this_sample_id,
  multiple_samples = FALSE,
  sample_list,
  from_flatfile = TRUE,
  projection = "grch37",
  with_chr_prefix = FALSE,
  streamlined = FALSE
)

Arguments

this_sample_id

Optional argument, single sample_id for the sample to retrieve segments for.

multiple_samples

Set to TRUE to return cn segments for multiple samples specified in samples_list parameter. Default is FALSE.

sample_list

Optional vector of type character with one sample per row, required if multiple_samples is set to TRUE.

from_flatfile

Set to TRUE by default.

projection

Selected genome projection for returned CN segments. Default is "grch37".

with_chr_prefix

Set to TRUE to add a chr prefix to chromosome names. Default is FALSE.

streamlined

Return a minimal output rather than full details. Default is FALSE.

Value

A data frame of segments for a specific or multiple sample ID(s).

Details

This function returns CN segments for samples. This works for single sample or multiple samples. For multiple samples, remember to set the Boolean parameter multiple_samples = TRUE and give the sample_lsit a vector of characters with one sample ID per row. For more information on how this function can be run in different ways, refer to the parameter descriptions, examples and vignettes. Is this function not what you are looking for? Try one of the following, similar, functions; assign_cn_to_ssm, get_cn_segments, get_cn_states,

Examples

#Return cn segments for one sample:
sample_cn_seg = get_sample_cn_segments(this_sample_id = "HTMCP-01-06-00422-01A-01D",
                                       multiple_samples = FALSE)

#Return cn segments for multiple samples (provided as vector of sample IDs):
these_sample_list = c("00-15201_tumorA", "00-15201_tumorB")

samples = get_sample_cn_segments(multiple_samples = TRUE, 
                                 sample_list = these_sample_list)