Get SSM By Regions.

Efficiently retrieve all mutations across a range of genomic regions.

get_ssm_by_regions(
  regions_list,
  regions_bed,
  streamlined = TRUE,
  maf_data = maf_data,
  use_name_column = FALSE,
  from_indexed_flatfile = TRUE,
  mode = "slms-3",
  augmented = TRUE,
  seq_type = "genome",
  projection = "grch37",
  min_read_support = 4,
  basic_columns = FALSE
)

Arguments

regions_list

Either provide a vector of regions in the chr:start-end format OR.

regions_bed

Better yet, provide a bed file with the coordinates you want to retrieve.

streamlined

Return a basic rather than full MAF format, default is TRUE.

maf_data

Use an already loaded MAF data frame.

use_name_column

If your bed-format data frame has a name column (must be named "name") these can be used to name your regions.

from_indexed_flatfile

Set to TRUE to avoid using the database and instead rely on flatfiles (only works for streamlined data, not full MAF details).

mode

Only works with indexed flatfiles. Accepts 2 options of "slms-3" and "strelka2" to indicate which variant caller to use. Default is "slms-3".

augmented

default: TRUE. Set to FALSE if you instead want the original MAF from each sample for multi-sample patients instead of the augmented MAF

seq_type

The seq_type you want back, default is genome.

projection

Obtain variants projected to this reference (one of grch37 or hg38).

min_read_support

Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs).

basic_columns

Boolean parameter set to FALSE per default. Set to TRUE to return fewer columns.

Value

Returns a data frame of variants in MAF-like format.

Details

This function internally calls get_ssm_by_region to retrieve SSM calls for the specified regions. See parameter descriptions for get_ssm_by_region for more information on how the different parameters can be called. Is this function not what you are looking for? Try one of the following, similar, functions; get_coding_ssm, get_coding_ssm_status, get_ssm_by_patients, get_ssm_by_sample, get_ssm_by_samples, get_ssm_by_region

Examples

#basic usage, adding custom names from bundled ashm data frame
regions_bed = dplyr::mutate(grch37_ashm_regions, name = paste(gene, region, sep = "_"))
#> Error in dplyr::mutate(grch37_ashm_regions, name = paste(gene, region,     sep = "_")): object 'grch37_ashm_regions' not found

ashm_basic_details = get_ssm_by_regions(regions_bed = regions_bed)
#> Error in apply(regions_bed, 1, bed2region): object 'regions_bed' not found

full_details_maf = get_ssm_by_regions(regions_bed = regions_bed,
                                      basic_columns=T)
#> Error in apply(regions_bed, 1, bed2region): object 'regions_bed' not found