Get SSM By Sample.

Get the SSMs (i.e. load MAF) for a single sample.

get_ssm_by_sample(
  this_sample_id,
  this_seq_type,
  these_samples_metadata,
  tool_name = "slms-3",
  projection = "grch37",
  these_genes,
  augmented = TRUE,
  flavour = "clustered",
  min_read_support = 3,
  basic_columns = TRUE,
  maf_cols = NULL,
  verbose = FALSE
)

Arguments

this_sample_id

Required. The sample_id you want the data from.

this_seq_type

Required if not specifying these_samples_metadata. The seq_type of the sample you want data from.

these_samples_metadata

Required if not specifying both this_sample_id and this_seq_type a single row or entire metadata table containing your sample_id.

tool_name

The name of the variant calling pipeline (currently only slms-3 is supported).

projection

The projection genome build. Supports hg38 and grch37.

these_genes

A vector of genes to subset ssm to.

augmented

default: TRUE. Set to FALSE if you instead want the original MAF from each sample for multi-sample patients instead of the augmented MAF.

flavour

Currently this function only supports one flavour option but this feature is meant for eventual compatibility with additional variant calling parameters/versions.

min_read_support

Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs).

basic_columns

Return first 43 columns of MAF rather than full details. Default is TRUE.

maf_cols

if basic_columns is set to FALSE, the user can specify what columns to be returned within the MAF. This parameter can either be a vector of indexes (integer) or a vector of characters.

verbose

Enable for debugging/noisier output.

Value

data frame in MAF format.

Details

This was implemented to allow flexibility because there are some samples that we may want to use a different set of variants than those in the main GAMBL merge. The current use case is to allow a force_unmatched output to be used to replace the SSMs from the merge for samples with known contamination in the normal. This will also be useful to apply a blacklist to individual MAFs when coupled with annotate_ssm_blacklist. Is this function not what you are looking for? Try one of the following, similar, functions; get_coding_ssm, get_coding_ssm_status, get_ssm_by_patients, get_ssm_by_samples, get_ssm_by_region, get_ssm_by_regions

Examples

this_sample_df = get_ssm_by_sample(this_sample_id = "HTMCP-01-06-00485-01A-01D",
                                   this_seq_type = "genome",
                                   tool_name = "slims-3",
                                   projection = "grch37")

capture_meta = get_gambl_metadata(seq_type_filter = "capture")

ssm_sample = get_ssm_by_sample(this_sample_id = "CASA0002_2015-03-10",
                               projection = "grch37",
                               augmented = T,
                               these_samples_metadata = capture_meta)
#> Warning: Column name 'GENE_PHENO' not found in column name header (case sensitive), skipping.
#> Warning: Column name 'FILTER' not found in column name header (case sensitive), skipping.
#> Warning: Column name 'flanking_bps' not found in column name header (case sensitive), skipping.
#> Warning: Column name 'vcf_id' not found in column name header (case sensitive), skipping.
#> Warning: Column name 'vcf_qual' not found in column name header (case sensitive), skipping.
#> Warning: Column name 'gnomAD_AF' not found in column name header (case sensitive), skipping.
#> Warning: Column name 'gnomAD_AFR_AF' not found in column name header (case sensitive), skipping.
#> Warning: Column name 'gnomAD_AMR_AF' not found in column name header (case sensitive), skipping.
#> Warning: Column name 'gnomAD_SAS_AF' not found in column name header (case sensitive), skipping.
#> Warning: Column name 'vcf_pos' not found in column name header (case sensitive), skipping.
#> Warning: Column name 'gnomADg_AF' not found in column name header (case sensitive), skipping.
#> Warning: Column name 'blacklist_count' not found in column name header (case sensitive), skipping.
#> Warning: Column name 'FILTER' (colClasses[[1]][67]) not found
#> Warning: Column name 'flanking_bps' (colClasses[[1]][68]) not found
#> Warning: Column name 'vcf_id' (colClasses[[1]][69]) not found
#> Warning: Column name 'vcf_qual' (colClasses[[1]][70]) not found
#> Warning: Column name 'gnomADg_AF' (colClasses[[1]][71]) not found
#> Warning: Column name 'vcf_pos' (colClasses[[2]][11]) not found
#> Warning: Column name 'gnomAD_AF' (colClasses[[3]][20]) not found
#> Warning: Column name 'gnomAD_AFR_AF' (colClasses[[3]][21]) not found
#> Warning: Column name 'gnomAD_AMR_AF' (colClasses[[3]][22]) not found
#> Warning: Column name 'gnomAD_SAS_AF' (colClasses[[3]][28]) not found
#> Warning: Column name 'GENE_PHENO' (colClasses[[4]][5]) not found
#> Warning: Column name 'blacklist_count' (colClasses[[4]][6]) not found