Liftover Bedpe.

Use liftOver to convert a bedpe file between the two main genome builds (grch37/hg38).

liftover_bedpe(
  bedpe_file,
  bedpe_df,
  target_build = "grch37",
  col_names,
  col_types,
  standard_bed = FALSE,
  verbose = FALSE
)

Arguments

bedpe_file

Either specify the path to a bedpe file.

bedpe_df

Or specify the bedpe data in a data frame.

target_build

Specify which build the data should be lifted to (must be one of hg19, grch37, hg38, grch38).

col_names

If not provided, the column names will be imposed.

col_types

Specify column types if column names are also defined with col_names.

standard_bed

Boolean parameter for defining the type of bed file that is provided with bedpe_file. Deafult is FALSE.

verbose

Set to TRUE for verbose output. Default is FALSE.

Value

Data frame containing original bedpe data with new coordinates.

Details

The user can specify a path to the bedpe file that needs to be lifted with bedpe_file, or, the suer can specify the bedpe data in a data frame with bedpe_df. The other required parameter is target_build, this parameter decides the final projection of the lifted bedpe file.

Examples

hg19_sv = get_manta_sv(verbose = FALSE)
#> WARNING! No SV calls found in flat-file for: 171116-PL02
#> WARNING! No SV calls found in flat-file for: 171447-PL01
#> WARNING! No SV calls found in flat-file for: 171451-PL01
hg19_sv = head(hg19_sv, 100)

hg38_sv = liftover_bedpe(bedpe_df = hg19_sv, 
                         target_build = "hg38")
#> Discarding unchained sequences: chrhs37d5