Maf To Custom Track.

Convert a maf-formatted data frame into a bed custom track file for UCSC.

maf_to_custom_track(
  maf_data,
  these_samples_metadata,
  seq_type = "genome",
  output_file,
  as_bigbed = FALSE,
  colour_column = "lymphgen",
  as_biglolly = FALSE,
  track_name = "GAMBL mutations",
  track_description = "mutations from GAMBL",
  verbose = FALSE,
  padding_size = 0
)

Arguments

maf_data

Either a maf loaded from disk or from the database using a get_ssm function.

these_samples_metadata

Optional argument, a metadata table subset to the samples of interest. If not provided, the function will run get_gambl_metadata for all available samples.

seq_type

The seq type you want back, default is "genome".

output_file

Name for your new bed file that can be uploaded as a custom track to UCSC.

as_bigbed

Boolean parameter controlling the format of the returned file. Default is FALSE.

colour_column

Set the colouring properties of the returned bed file. Per default, this function will assign colour based on "lymphgen".

as_biglolly

Boolean parameter controlling the format of the returned file. Default is FALSE (i.e a BED file will be returned).

track_name

Track name. Default is "GAMBL mutations"

track_description

Track description. Default is "mutations from GAMBL"

verbose

Default is FALSE.

padding_size

Optional parameter specifying the padding size in the returned file, default is 0.

Value

Nothing.

Details

This function takes an incoming MAF and converts it to a UCSC Genome Browser ready BED (or bigbed/biglolly) file. Optional parameters available for further customization of the returned file. For more information, refer to the parameter descriptions and function examples.

Examples

if (FALSE) {
maf_to_custom_track(my_maf_data, "/home/rmorin/private/some_mutations.bed")
}