Make IGV Snapshot

Load bams and generate an IGV screenshot for one or more regions.

make_igv_snapshot(
  bams,
  genome_build,
  region,
  padding = 200,
  chrom,
  start,
  end,
  this_sample_id,
  out_path = "/tmp/",
  igv_port = 60506
)

Arguments

bams

Character vector containing the full path to one or more bam files.

genome_build

String specifying the genome build for the bam files provided.

region

Optionally specify the region as a single string (e.g. "chr1:1234-1235").

padding

Optionally specify a positive value to broaden the region around the specified position. Default is 200.

chrom

Optionally specify the region by specifying the chromosome, start and end (see below).

start

Optionally specify the region by specifying the start.

end

Optionally specify the region by specifying the end.

this_sample_id

Specify the sample_id or any other string you want embedded in the file name.

out_path

Specify the output directory where the snapshot will be written.

igv_port

Specify the port IGV is listening on.

Value

Path to file (.png).

Details

Specify the path to one or more bam files as a character vector to the bams parameter. The user can also specify regions of interest with either the region parameter (chr:start-end), or the user can directly supply the chromosome, start and end coordinates with the chrom, start, and end parameters. For more information and examples, refer to the function examples and parameter descriptions.

Examples