INTERNAL FUNCTION called by plot_sample_circos and get_mutation_frequency_bin_matrix, not meant for out-of-package usage. Assign a colour palette to metadata columns automatically and consistently.

INTERNAL FUNCTION called by plot_sample_circos and get_mutation_frequency_bin_matrix, not meant for out-of-package usage. Assign a colour palette to metadata columns automatically and consistently.

map_metadata_to_colours(
  metadataColumns,
  these_samples_metadata,
  column_alias = list(),
  as_vector = TRUE,
  verbose = FALSE,
  annoAlpha = 1
)

Arguments

metadataColumns

Names of the metadata columns to assign colours for.

these_samples_metadata

Metadata for just the samples you need colours for.

column_alias

A list of column_names with values indicating what gambl colour scheme they should use (e.g. pos_neg, pathology, lymphgen).

as_vector

Boolean statement that is set to TRUE per default.

verbose

Set to TRUE to enable verbose mode (debugging messages.

annoAlpha

Optional alpha to apply to annotation colours.

Value

Either a vector or list of colours.

Examples

#get metadata
all_meta = get_gambl_metadata()

#get colours
all_cols = map_metadata_to_colours(metadataColumns = c("lymphgen",
                                                       "pathology",
                                                       "genetic_subgroup"),
                                   these_samples_metadata = all_meta,
                                   column_alias = list("nothing" = "FL"),
                                   as_vector = F)
#> Error in map_metadata_to_colours(metadataColumns = c("lymphgen", "pathology",     "genetic_subgroup"), these_samples_metadata = all_meta, column_alias = list(nothing = "FL"),     as_vector = F): could not find function "map_metadata_to_colours"