VAF Heatmap

Plot a heatmap comparing the VAF of mutations in T1/T2 pairs.

plot_mutation_dynamics_heatmap(
  maf1,
  maf2,
  vafcolname,
  patient_colname = "patient_id",
  these_samples_metadata,
  sortByColumns,
  metadata_columns = c("sample_id"),
  gene_orientation = "bottom",
  annotate_zero = FALSE,
  genes,
  top_n_genes,
  drop_unless_lowvaf = FALSE,
  vaf_cutoff_to_drop = 0.05,
  cluster_columns = FALSE,
  cluster_rows = FALSE
)

Arguments

maf1

Data frame of simple somatic mutations at time point A.

maf2

Data frame of simple somatic mutations at time point B.

vafcolname

Name of variable that holds VAF in maf. If not supplied, vaf will be calcualted.

patient_colname

Column name that holds patient name (default is "patient_id").

these_samples_metadata

GAMBL metadata subset to the cases you want to process (or full metadata).

sortByColumns

Which of the metadata to sort on for the heatmap.

metadata_columns

A vector containing the categorical column names you want to plot below.

gene_orientation

Where genes would be plotted. Default is "bottom".

annotate_zero

Indicate a variant that had VAF = 0 in one of the two time points. Default is FALSE.

genes

An optional vector of genes to restrict your plot to.

top_n_genes

How many genes to be added to the plot.

drop_unless_lowvaf

Will drop some genes where VAF is low, default is FALSE.

vaf_cutoff_to_drop

Which VAF cut-off value to use when dropping variants before plotting.

cluster_columns

Boolean statement for clustering by columns, defaults to FALSE.

cluster_rows

Boolean statement for clustering by rows, defaults to FALSE.

Value

Nothing

Details

Currently unfinished plotting function. Thus, I have removed it from export until it's in a state where it can be included in GAMBLR. Parameter descriptions need to be updated so that the origin of the incoming data is clear. Examples would also need to be added before this function gets exported into NAMESPACE.