Populate Each Tool Results.

Convenience function for returning results from a specified tool.

populate_each_tool_result(
  tool,
  genome_builds,
  unix_groups,
  include_silent = FALSE
)

Arguments

tool

Name of tool to get results from.

genome_builds

A single genome build or a vector of all genome builds to process.

unix_groups

A single unix group or a vector of all unix groups to process.

include_silent

Logical parameter indicating whether to include silent mutations into coding mutations. Default is FALSE.

Value

Nothing.

Details

This function takes a tool name tool as well as other parameters for specifying the requested result. Other parameters include genome_build, this can be just one parameter or a vector with different genome builds to return results for. Similarly, unix_group can take either one value or a vector with all the different unix groups to return results for. Lastly, the user can subset the returned results to only silent mutations. This is done with setting include_silent = TRUE (default is FALSE).

Examples

tool_results = populate_each_tool_result("smlims_3", "grch37", "gambl", FALSE)
#> [1] "missing:  /Users/rmorin/git/gambl/data/metadata/gambl_all_outcomes.tsv"
#> Have you cloned the GAMBL repo and added the path to this directory under the local section of your config?
#> Error: '/Users/rmorin/git/gambl/data/metadata/gambl_all_outcomes.tsv' does not exist.