Gene Cloud Plot

Make a word cloud of gene names from a MAF file based on mutation frequency.

prettyGeneCloud(
  maf_df,
  these_genes,
  other_genes,
  these_genes_colour = "#B2DF8A",
  other_genes_colour = "#bc42f5",
  colour_index
)

Arguments

maf_df

A MAF-format data frame containing the mutations you want to summarize in a gene word cloud

these_genes

An optional vector of gene symbols (defaults to all lymphoma genes)

other_genes

An optional second vector of gene symbols to include in your cloud in a second colour

these_genes_colour

Specify the hex code of a colour to use for the first set of genes

other_genes_colour

Specify another hex code of a colour to use for the second set of genes

colour_index

Optional named character vector with a name for each gene in these_genes and a colour as the value

Value

data frame with counts for each gene

Details

Create a wordcloud from an incoming MAF. Required parameter is maf_df. Optional parameters are these_genes, other_genes, these_genes_colour, other_genes_colour and colour_index. If no genes are supplied when calling the function, this function will default to all lymphoma genes.

Examples

#get all coding SSM
maf = get_coding_ssm(seq_type = "genome")
#> reading from: /projects/nhl_meta_analysis_scratch/gambl/results_local/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/augmented_maf/all_slms-3--grch37.CDS.maf
#> mutations from 1652 samples
#> after linking with metadata, we have mutations from 1646 samples

#get gene symbols from MAF
maf_genes = dplyr::filter(maf, Hugo_Symbol != "Unknown")
maf_genes_chr1 = dplyr::filter(maf_genes, Chromosome == "1")
my_genes = pull(maf_genes_chr1, Hugo_Symbol)
#> Error in pull(maf_genes_chr1, Hugo_Symbol): could not find function "pull"

#build wordcloud
prettyGeneCloud(maf_df = maf, these_genes = my_genes)
#> Error in dplyr::filter(maf_df, Hugo_Symbol %in% these_genes): Problem while computing `..1 = Hugo_Symbol %in% these_genes`.
#> Caused by error in `h()`:
#> ! error in evaluating the argument 'table' in selecting a method for function '%in%': object 'my_genes' not found