Make an oncoplot that is pretty using ComplexHeatmap. The metadata is expected to follow the structure and column naming used in GAMBL. If you provide your own non-GAMBL samples and metadata, you must include at least the following columns with these names. The first one should match the Tumor_Sample_Barcode in the MAF object or onco_matrix you provide. sample_id, pathology

Make an oncoplot that is pretty using ComplexHeatmap. The metadata is expected to follow the structure and column naming used in GAMBL. If you provide your own non-GAMBL samples and metadata, you must include at least the following columns with these names. The first one should match the Tumor_Sample_Barcode in the MAF object or onco_matrix you provide. sample_id, pathology

prettyOncoplot(
  maftools_obj,
  onco_matrix_path,
  genes,
  keepGeneOrder = TRUE,
  keepSampleOrder = TRUE,
  highlightHotspots = FALSE,
  these_samples_metadata,
  metadataColumns,
  numericMetadataColumns,
  expressionColumns = c(),
  numericMetadataMax,
  sortByColumns,
  removeNonMutated = FALSE,
  minMutationPercent,
  fontSizeGene = 6,
  annoAlpha = 1,
  mutAlpha = 1,
  recycleOncomatrix = FALSE,
  box_col = NA,
  metadataBarHeight = 1.5,
  metadataBarFontsize = 5,
  hideTopBarplot = FALSE,
  hideSideBarplot = FALSE,
  splitColumnName,
  splitGeneGroups,
  legend_row = 3,
  legend_col = 3
)

Arguments

maftools_obj
A maftools object containing the mutations you want to plot

onco_matrix_path
Provide a path to an onco_matrix file instead of a MAF object if the former is unavailable (this limits functionality a bit)

genes
An optional list of genes to restrict your plot to

keepGeneOrder
Set to TRUE if you want to preserve the gene order specified

keepSampleOrder
Set to TRUE if you want to preserve the sample order specified

these_samples_metadata
Data frame containing metadata for your samples

metadataColumns
A vector containing the categorical column names you want to plot below

numericMetadataColumns
A vector containing the numeric columns you want to plot below

numericMetadataMax
A numeric vector of cutoffs to apply to numeric columns above

sortByColumns
A vector containing the column names you want to sort columns (patients) on

removeNonMutated
Set to TRUE to drop unmutated cases

minMutationPercent
Only genes mutated in more than minMutationPercent % patients will be included

recycleOncomatrix
Set to TRUE most of the time to reuse the oncomatrix saved by maftools

legend_row
Fiddle with these to widen or narrow your legend

legend_col
Fiddle with these to widen or narrow your legend

Value