Rainfall Plot

Plot a rainfall plot for one sample. This function takes in the MAF data frame, or path to a custom MAF file. If non are specified, the SSM will be obtained through GAMBLR directly (with get_ssm_by_region).

prettyRainfallPlot(
  this_sample_id,
  label_ashm_genes = TRUE,
  projection = "grch37",
  chromosome,
  this_maf,
  maf_path,
  zoom_in_region,
  seq_type,
  label_sv = FALSE
)

Arguments

this_sample_id

Sample id for the sample to display. This is argument is not required if you want a multi-sample plot but is otherwise needed.

label_ashm_genes

Boolean argument indicating whether the aSHM regions will be labeled or not.

projection

Specify projection (grch37 or hg38) of mutations. Default is grch37.

chromosome

Provide one or more chromosomes to plot. The chr prefix can be inconsistent with projection and will be handled.

this_maf

Specify custom MAF data frame of mutations.

maf_path

Specify path to MAF file if it is not already loaded into data frame.

zoom_in_region

Provide a specific region in the format "chromosome:start-end" to zoom in to a specific region.

seq_type

Specify one of "genome" or "capture" when relying on the function to obtain mutations from a region (i.e. if you haven't provided a MAF or single sample_id)

label_sv

Boolean argument to specify whether label SVs or not. Only supported if a specific chromosome or zoom in region are specified.

Value

a ggplot2 plot. Print it using print() or save it using ggsave()

Details

Create a sample-level rainfall plot visualizing single nucleotide substitutions mutations for selected chromosomes.

Examples

prettyRainfallPlot(this_sample_id = "Raji",
                   seq_type = "genome",
                   zoom_in_region = "8:125252796-135253201",
                   label_sv = TRUE)
#> Error in dplyr::rename(., start = "hg19_start", end = "hg19_end", Chromosome = "chr_name"): Can't rename columns that don't exist.
#> ✖ Column `hg19_start` doesn't exist.

#multi-sample rainfall plot for one gene region
prettyRainfallPlot(zoom_in_region = "chr3:5,221,286-5,269,723",
                   seq_type = "genome")
#> Warning: No sample_id was provided. Using all mutations in the MAF within your region!
#> Error in dplyr::rename(., start = "hg19_start", end = "hg19_end", Chromosome = "chr_name"): Can't rename columns that don't exist.
#> ✖ Column `hg19_start` doesn't exist.