Review Hotspots.

Annotate MAF-like data frome with a hot_spot column indicating recurrent mutations.

review_hotspots(
  annotated_maf,
  genes_of_interest = c("FOXO1", "MYD88", "CREBBP", "NOTCH1", "NOTCH2", "CD79B", "EZH2"),
  genome_build = "grch37"
)

Arguments

annotated_maf

A data frame in MAF format that has hotspots annotated using the function annotate_hotspots().

genes_of_interest

A vector of genes for hotspot review. Currently only FOXO1, MYD88, CREBBP, NOTCH1, NOTCH2, CD79B and EZH2 are supported.

genome_build

Reference genome build for the coordinates in the MAF file. The default is grch37 genome build.

Value

The same data frame (as given to the annotated_maf parameter) with the reviewed column "hot_spot".

Details

This function takes an annotated MAF (with annotate_hotspots) and adds a new column, "hot_spot", to the same data frame. Genes for hotspot review are supplied with the genes_of_interest parameter. Currently only a few sets of genes are supported, see parameter description for more information and limitations. The desired genome build can be specified with genome_build parameter. Should be the same as the incoming MAF.

Examples

hot_ssms = review_hotspots(annotate_hotspots(get_coding_ssm(seq_type = "genome")),
                           genes_of_interest = c("CREBBP"))
#> Adding missing grouping variables: `SYMBOL`
#> Adding missing grouping variables: `SYMBOL`
#> reading from: /projects/nhl_meta_analysis_scratch/gambl/results_local/all_the_things/slms_3-1.0_vcf2maf-1.3/genome--projection/deblacklisted/augmented_maf/all_slms-3--grch37.CDS.maf
#> mutations from 1652 samples
#> after linking with metadata, we have mutations from 1646 samples