Get ASHM Count Matrix.

Prepare a matrix with one row per sample and one column per region using a set of hypermutated regions.

get_ashm_count_matrix(
  regions_bed,
  maf_data,
  these_samples_metadata,
  seq_type,
  from_indexed_flatfile = TRUE
)

Arguments

regions_bed

A bed file with one row for each region.

maf_data

Optionally provide a data frame in the MAF format, otherwise the database will be used.

these_samples_metadata

This is used to complete your matrix. All GAMBL samples will be used by default. Provide a data frame with at least sample_id for all samples if you are using non-GAMBL data.

seq_type

The seq type to return results for.

from_indexed_flatfile

Boolean parameter set to TRUE per default.

Value

A matrix.

Details

Values are the number of mutations in that patient in the region.

Examples

regions_bed = dplyr::mutate(grch37_ashm_regions, name = paste(gene, region, sep = "_"))

matrix = get_ashm_count_matrix(regions_bed = regions_bed,
                               seq_type = "genome")
#>  [1] "chr2:96808901-96811913"    "chr17:56407732-56410140"  
#>  [3] "chr11:128339774-128345731" "chr11:128388492-128394163"
#>  [5] "chr6:31548325-31550717"    "chr3:32020518-32024930"   
#>  [7] "chr11:65265237-65268359"   "chr11:65190369-65192380"  
#>  [9] "chr16:11347512-11350007"   "chr1:226921088-226927982" 
#> [11] "chr19:16434978-16439011"   "chr9:37023396-37027663"   
#> [13] "chr9:37192080-37207549"    "chr9:37282092-37309161"   
#> [15] "chr9:37323113-37340687"    "chr9:37369209-37372160"   
#> [17] "chr9:37382267-37385854"    "chr9:37395932-37409239"   
#> [19] "chr9:37032576-37037704"    "chr3:71626341-71635648"   
#> [21] "chr3:71550753-71551478"    "chr3:71503103-71504870"   
#> [23] "chr3:71350633-71357665"    "chr3:71176427-71180960"   
#> [25] "chr3:71104986-71119518"    "chr3:101546669-101547704" 
#> [27] "chr3:101568239-101569274"  "chr3:186783199-186784291" 
#> [29] "chr3:186739628-186740875"  "chr18:60982728-60988342"  
#> [31] "chr18:60796984-60814103"   "chr3:187458526-187464632" 
#> [33] "chr12:122456912-122464036" "chr11:102188170-102190077"
#> [35] "chr8:11347723-11355318"    "chr12:92537999-92539598"  
#> [37] "chr1:203274698-203275778"  "chr6:14118026-14120025"   
#> [39] "chr16:10970795-10975465"   "chr2:136874728-136875461" 
#> [41] "chr12:113492311-113497546" "chr12:11796001-11812968"  
#> [43] "chr10:127578912-127591133" "chr6:390572-394093"       
#> [45] "chr16:85931918-85933977"   "chr8:128748352-128749427" 
#> [47] "chr20:46128611-46138099"   "chr9:37021323-37043181"   
#> [49] "chr6:37138104-37139804"    "chrX:48774756-48776255"   
#> [51] "chr19:10340142-10341764"   "chr14:94940587-94942549"  
#> [53] "chr3:186781067-186784816"  "chr3:186737482-186741455" 
#> [55] "chrX:12993308-12994511"    "chr9:37276495-37342191"   
#> [57] "chr14:69257848-69259739"   "chr4:40193105-40204231"   
#> [59] "chr19:19279635-19281441"   "chr6:90981034-91016134"   
#> [61] "chr13:46957278-46963342"   "chr5:158500476-158532769" 
#> [63] "chr22:23229554-23232042"   "chr15:59658991-59671152"  
#> [65] "chr3:16546433-16556786"    "chr14:96179535-96180366"  
#> [67] "chr2:60773789-60783486"    "chr2:111874854-111882174" 
#> [69] "chr11:35156769-35164248"   "chr5:149790977-149792349" 
#> [71] "chr9:37423010-37425279"    "chr16:27322895-27329423"  
#> [73] "chr5:131823933-131826458"  "chr1:226920563-226927885" 
#> [75] "chr1:226864857-226873452"  "chr3:183269360-183274139" 
#> [77] "chr3:187771678-187982852"  "chr3:188377178-188491248" 
#> [79] "chr12:25204045-25213569"   "chr5:88131209-88206620"   
#> [81] "chr21:26934372-26937651"   "chr11:60223385-60225310"  
#> [83] "chr17:3597616-3599572"     "chr11:111248078-111250224"
#> [85] "chr20:49120482-49140477"   "chr17:75424734-75471471"  
#> [87] "chr6:134487960-134499859"  "chr7:5568297-5570856"     
#> Joining, by = c("sample_id", "region_name")