Load bams and generate an IGV screenshot for one or more regions

Load bams and generate an IGV screenshot for one or more regions

make_igv_snapshot(
  bams,
  genome_build,
  region,
  padding = 200,
  chrom,
  start,
  end,
  sample_id,
  out_path = "/tmp/",
  igv_port = 60506
)

Arguments

bams
Character vector containing the full path to one or more bam files

genome_build
String specifying the genome build for the bam files provided

region
Optionally specify the region as a single string (e.g. "chr1:1234-1235")

padding
Optionally specify a positive value to broaden the region around the specified position

chrom
Optionally specify the region by specifying the chromosome, start and end (see below)

start
Optionally specify the region by specifying the start

end
Optionally specify the region by specifying the end

sample_id
Specify the sample_id or any other string you want embedded in the file name

out_path
Specify the output directory where the snapshot will be written

igv_port
Specify the port IGV is listening on

region_bed
Optionally specify regions in bed format (column order is assumed)

Value

Examples

#IMPORTANT: you must be running IGV on the host that is running R and you need to have it listening on a port
# The simplest scenario is to run this command on a terminal (if using a Mac), assuming you are using R on gphost10 and you have a ssh config that routes gp10 to that host
# ssh -X gp10
# then launch IGV (e.e. from a conda installation):
# conda activate igv; igv &
this_sv = annotated_sv %>% filter(gene=="ETV6")

#> Error in annotated_sv %>% filter(gene == "ETV6"): could not find function "%>%"

tumour_bam = get_bam_path(sample=this_sv$tumour_sample_id)

#> Error in get_bam_path(sample = this_sv$tumour_sample_id): could not find function "get_bam_path"

make_igv_snapshot(chrom=this_sv$chrom2, start=this_sv$start2, end=this_sv$end2, sample_id=this_sv$tumour_sample_id,out_path="~/IGV_snapshots/")

#> Socket connection # 19 to on port 60506 
#> [1] "new \n"

#> Error in paste0(chrom, ":", start - padding, "-", end + padding): object 'this_sv' not found